chore: fix test errors
